requirements.txt
# Main Dependencies
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
pydantic[email]==2.1.1
pymongo==4.5.0
motor==3.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
streamlit==1.24.0
requests==2.31.0
pandas==2.2.3
python-dotenv==1.0.0
groq==0.4.0
httpx==0.24.1
aiohttp==3.11.18
matplotlib==3.10.3